IIoT/pyproject.toml
[project]
name = "iiot"
version = "0.1.0"
description = ""
authors = [
    {name = "giicoo", email = "giicoo.gm@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "paho-mqtt (==1.6.1)",          # <-- зафиксировано на 1.6.1
    "asyncio-mqtt (>=0.16.2,<0.17.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "app" }
]
